Changed redirections to login

+++ src/hr_management/user/views.py
@@ -26,7 +26,7 @@ def admin_home(request):
 		context = {"title": "Dashboard"}
 		return render(request, "../templates/administrator/admin_home.html", context)
 	else:
-		messages.success(request, "(from admin_home) You must be Admin and be Logged In To View That Page...")
+		messages.success(request, "You must be logged in as admin to access admin dashboard.")
 		return redirect('login')  
 
 
@@ -62,8 +62,8 @@ def manage_employees(request):
         }
         return render(request, 'administrator/manage_employees.html', context)
     else:
-        messages.success(request, "(from manage_employees) You must be an Admin, be logged in, and have permission to view this page...")
-        return redirect('home')
+        messages.success(request, "You must be logged in as admin to manage company employees.")
+        return redirect('login')
 
 
 
@@ -85,8 +85,8 @@ def register_employee(request):
 
 		return render(request, '../templates/administrator/register_employee.html', {'form':form})
 	else:
-		messages.success(request, "(from register_employee) You Must Be Logged In To View That Page...")
-		return redirect('manage_employees')
+		messages.success(request, "You must be logged in as admin to register a new employee.")
+		return redirect('login')
 
 
 
@@ -95,8 +95,8 @@ def employee_record(request, pk):
 		employee_record = Employee.objects.get(id=pk) # Look Up Records
 		return render(request, '../templates/administrator/employee_record.html', {'title': 'Employee Record', 'employee_record':employee_record})
 	else:
-		messages.success(request, "You Must Be Logged In To View That Page...")
-		return redirect('home')
+		messages.success(request, "You must be logged in as admin to view employee records.")
+		return redirect('login')
 
 
 
@@ -117,20 +117,20 @@ def update_employee(request, pk):
         }
 		return render(request, '../templates/administrator/update_employee.html', context)
 	else:
-		messages.success(request, "You Must Be Logged In...")
-		return redirect('home')
+		messages.success(request, "You must be logged in as admin to update employee info.")
+		return redirect('login')
 	
 
 
 def delete_employee(request, pk):
-	if request.user.is_authenticated:
+	if request.user.is_authenticated and request.user.is_staff:
 		delete_it = Employee.objects.get(id=pk)
 		delete_it.delete()
 		messages.success(request, "Employee Deleted Successfully")
 		return redirect('manage_employees')
 	else:
-		messages.success(request, "You Must Be Logged In To Do That...")
-		return redirect('home')
+		messages.success(request, "You must be logged in as admin to delete an employee.")
+		return redirect('login')
 
 
 
@@ -154,7 +154,7 @@ def employee_home(request):
 		}
 		return render(request, '../templates/employees/employee_home.html', context)
 	else:
-		messages.success(request, "You must be Employee and be Logged In To View That Page...")
+		messages.success(request, "You must be logged in as employee to view the employee home page.")
 		return redirect('login')  
 
 
@@ -171,8 +171,8 @@ def edit_profile(request):
 			return redirect('employee_home')
 		return render(request, '../templates/employees/edit_profile.html', {'title': 'Edit My Profile', 'form':form})
 	else:
-		messages.success(request, "You Must Be Logged In...")
-		return redirect('home')
+		messages.success(request, "You must be logged in as employee to access the edit profile page.")
+		return redirect('login')
 
 
 # ======> Help Views <======
@@ -182,5 +182,5 @@ def help(request):
 	elif request.user.is_authenticated and not(request.user.is_staff):
 		return render(request, '../templates/help/employee_home_help.html', {'title': 'Help'})
 	else:
-		messages.success(request, "You Must Be Logged In...")
-		return redirect('home')
+		messages.success(request, "You must be logged in to view the help page.")
+		return redirect('login')
